Fixes/booking

+++ public/assets/js/payment.js
@@ -18,38 +18,38 @@ function incrementSeconds() {
 let cancel = setInterval(incrementSeconds, 1000);
 
 setTimeout(delBooking, 602000);
-//TODO
-var show_close_alert=true;
-//window.onbeforeunload = delBooking;
-$("form").bind("submit", function (){
-    show_close_alert=false;
-    
-});
-$(window).bind("beforeunload", function () {
-    if(show_close_alert){
-        delBooking();
-    }
-    
-});
+
+let submitted = false;
+
+function setSubmitted(){
+    console.log(1);
+    submitted = true;
+}
+window.onbeforeunload = delBooking;
+
 
 function delBooking(){
-    $.ajax({
-        type: "DELETE",
-        url: '/booking/deleteBooking',
-        data: {
-            'booking_id' : $('#booking_id').val(),
+    if (!submitted) {
+        submitted = true;
+        console.log("Deleting booking")
+        $.ajax({
+            type: "DELETE",
+            url: '/booking/deleteBooking',
+            data: {
+                'booking_id': $('#booking_id').val(),
 
-        },
-        contentType: "application/x-www-form-urlencoded; charset=UTF-8",
-        success: function (response) {
-            if (response.result === 'redirect') {
-                //redirecting
-                let baseurl = window.location.origin;
-                let url = baseurl + response.url;
-                window.location.href = url;
+            },
+            contentType: "application/x-www-form-urlencoded; charset=UTF-8",
+            success: function (response) {
+                // if (response.result === 'redirect') {
+                //     //redirecting
+                //     let baseurl = window.location.origin;
+                //     let url = baseurl + response.url;
+                //     window.location.href = url;
+                // }
+            },
+            error: function (res) {
             }
-        },
-        error: function (res) {
-        }
-    });
+        });
+    }
 }
